Initial bonding contract

+++ casper/src/main/scala/coop/rchain/casper/Estimator.scala
@@ -18,7 +18,7 @@ object Estimator {
   type BlockHash = ByteString
   type Validator = ByteString
 
-  implicit val decreasingOrder = Ordering[Int].reverse
+  implicit val decreasingOrder = Ordering[Long].reverse
 
   /**
     * When the BlockDag has an empty latestMessages, tips will return IndexedSeq(genesis)
@@ -31,12 +31,12 @@ object Estimator {
     def sortChildren(
         blocks: IndexedSeq[BlockHash],
         childMap: Map[BlockHash, Set[BlockHash]],
-        scores: Map[BlockHash, Int]
+        scores: Map[BlockHash, Long]
     ): IndexedSeq[BlockHash] = {
       // TODO: This ListContrib.sortBy will be improved on Thursday with Pawels help
       val newBlocks =
         ListContrib
-          .sortBy[BlockHash, Int](
+          .sortBy[BlockHash, Long](
             blocks.flatMap(replaceBlockHashWithChildren(childMap, _, scores)).distinct,
             scores
           )
@@ -55,7 +55,7 @@ object Estimator {
     def replaceBlockHashWithChildren(
         childMap: Map[BlockHash, Set[BlockHash]],
         b: BlockHash,
-        scores: Map[BlockHash, Int]
+        scores: Map[BlockHash, Long]
     ): IndexedSeq[BlockHash] = {
       val c: Set[BlockHash] = childMap.getOrElse(b, Set.empty[BlockHash]).filter(scores.contains)
       if (c.nonEmpty) {
@@ -77,17 +77,17 @@ object Estimator {
   }
 
   // TODO: Fix to stop at genesis/LFB
-  def buildScoresMap[F[_]: Monad: BlockStore](blockDag: BlockDag): F[Map[BlockHash, Int]] = {
+  def buildScoresMap[F[_]: Monad: BlockStore](blockDag: BlockDag): F[Map[BlockHash, Long]] = {
     def hashParents(hash: BlockHash): F[List[BlockHash]] =
       for {
         b <- unsafeGetBlock[F](hash)
       } yield parentHashes(b).toList
 
     def addValidatorWeightDownSupportingChain(
-        scoreMap: Map[BlockHash, Int],
+        scoreMap: Map[BlockHash, Long],
         validator: Validator,
         latestBlockHash: BlockHash
-    ): F[Map[BlockHash, Int]] =
+    ): F[Map[BlockHash, Long]] =
       for {
         updatedScoreMap <- DagOperations
                             .bfTraverseF[F, BlockHash](List(latestBlockHash))(hashParents)
@@ -95,7 +95,7 @@ object Estimator {
                               case (acc, hash) =>
                                 for {
                                   b               <- unsafeGetBlock[F](hash)
-                                  currScore       = acc.getOrElse(hash, 0)
+                                  currScore       = acc.getOrElse(hash, 0L)
                                   validatorWeight <- weightFromValidator[F](b, validator)
                                 } yield acc.updated(hash, currScore + validatorWeight)
                             }
@@ -108,7 +108,7 @@ object Estimator {
       * TODO: Add test where this matters
       */
     def addValidatorWeightToImplicitlySupported(
-        scoreMap: Map[BlockHash, Int],
+        scoreMap: Map[BlockHash, Long],
         childMap: Map[BlockHash, Set[BlockHash]],
         validator: Validator,
         latestBlockHash: BlockHash
@@ -123,8 +123,8 @@ object Estimator {
                 for {
                   c <- ProtoUtil.unsafeGetBlock[F](cHash)
                   result = if (ProtoUtil.parentHashes(c).size > 1 && c.sender != validator) {
-                    val currScore       = acc2.getOrElse(cHash, 0)
-                    val validatorWeight = ProtoUtil.weightMap(c).getOrElse(validator, 0)
+                    val currScore       = acc2.getOrElse(cHash, 0L)
+                    val validatorWeight = ProtoUtil.weightMap(c).getOrElse(validator, 0L)
                     acc2.updated(cHash, currScore + validatorWeight)
                   } else {
                     acc2
@@ -134,22 +134,23 @@ object Estimator {
         }
 
     for {
-      scoresMap <- Foldable[List].foldM(blockDag.latestMessages.toList, Map.empty[BlockHash, Int]) {
-                    case (acc, (validator: Validator, latestBlock: BlockMessage)) =>
-                      for {
-                        postValidatorWeightScoreMap <- addValidatorWeightDownSupportingChain(
-                                                        acc,
-                                                        validator,
-                                                        latestBlock.blockHash
-                                                      )
-                        postImplicitlySupportedScoreMap <- addValidatorWeightToImplicitlySupported(
-                                                            postValidatorWeightScoreMap,
-                                                            blockDag.childMap,
-                                                            validator,
-                                                            latestBlock.blockHash
-                                                          )
-                      } yield postImplicitlySupportedScoreMap
-                  }
+      scoresMap <- Foldable[List]
+                    .foldM(blockDag.latestMessages.toList, Map.empty[BlockHash, Long]) {
+                      case (acc, (validator: Validator, latestBlock: BlockMessage)) =>
+                        for {
+                          postValidatorWeightScoreMap <- addValidatorWeightDownSupportingChain(
+                                                          acc,
+                                                          validator,
+                                                          latestBlock.blockHash
+                                                        )
+                          postImplicitlySupportedScoreMap <- addValidatorWeightToImplicitlySupported(
+                                                              postValidatorWeightScoreMap,
+                                                              blockDag.childMap,
+                                                              validator,
+                                                              latestBlock.blockHash
+                                                            )
+                        } yield postImplicitlySupportedScoreMap
+                    }
     } yield scoresMap
   }
 }
