+++ casper/src/test/scala/coop/rchain/casper/api/CreateBlockAPITest.scala
@@ -78,7 +78,7 @@ private class SleepingMultiParentCasperImpl[F[_]: Monad: Timer](underlying: Mult
   def deploy(d: DeployData): F[Either[Throwable, Unit]]     = underlying.deploy(d)
   def estimator(dag: BlockDag): F[IndexedSeq[BlockMessage]] = underlying.estimator(dag)
   def blockDag: F[BlockDag]                                 = underlying.blockDag
-  def normalizedInitialFault(weights: Map[Validator, Int]): F[Float] =
+  def normalizedInitialFault(weights: Map[Validator, Long]): F[Float] =
     underlying.normalizedInitialFault(weights)
   def lastFinalizedBlock: F[BlockMessage]          = underlying.lastFinalizedBlock
   def storageContents(hash: ByteString): F[String] = underlying.storageContents(hash)
